Completed logic for pills_fit_lut()
Calls pills_model_process to get sheet holding coefficents and offsets

Goes through the depth and fan angle values.

Use the fan angle as an index to get the coefficent and offset

Get THU and TVU at this point

Add to list of THU and TVU values.

Return result THU and TVU as numpy arrays