Completed hawkeye_fit_lut().
Get tvu, thu, and range bias coefficents and offsets from LUTs using hawkeye_model_process().

Calculate and return range bias uncertainty component.

If the point is not-subaqueous, subaqueous thu, subaqueous tvu, and range bias for that point is set to 0.

Depending on scanner_channel and user_data values, get coefficients from the correct LUT and calculate subaqueous thu and tvu for that point.